% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.functions.r
\name{gaussian.kernel}
\alias{gaussian.kernel}
\title{Gaussian kernel}
\usage{
gaussian.kernel(d, bandwidth)
}
\arguments{
\item{d}{Numeric vector of distances.}

\item{bandwidth}{Numeric bandwidth.}
}
\value{
A numeric vector of kernel weights.
}
\description{
Gaussian kernel function - weights all points by distance on a Gaussian curve with standard deviation of bandwidth.
}
\keyword{internal}
