% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.functions.r
\name{gaussian.square.kernel}
\alias{gaussian.square.kernel}
\title{Gaussian square kernel}
\usage{
gaussian.square.kernel(d, bandwidth)
}
\arguments{
\item{d}{Numeric vector of distances.}

\item{bandwidth}{Numeric bandwidth.}
}
\value{
A numeric vector of kernel weights.
}
\description{
Gaussian square kernel function - weights all points within bandwidth by distance on a Gaussian curve with standard deviation of bandwidth, but all points beyond bandwidth at 0.
}
\keyword{internal}
