% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d.smooths.r
\name{load.kernel.smooths}
\alias{load.kernel.smooths}
\title{Load kernel smooths}
\usage{
load.kernel.smooths(location)
}
\arguments{
\item{location}{String location to which results were saved.}
}
\value{
A list containing the parameters and a data.frame with the smoothed estimates (same structure as returned by kernel.smooth.in.space.and.time() and kernel.smooth.in.space.and.time.with.margins()).
}
\description{
Loads the output of kernel.smooth.in.space.and.time() or kernel.smooth.in.space.and.time.with.margins() previously saved with save.kernel.smooths()
}
