% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.functions.r
\name{sample.size.floored.variance}
\alias{sample.size.floored.variance}
\title{Sample size floored variance}
\usage{
sample.size.floored.variance(alpha, margin, dependent.variable, weights)
}
\arguments{
\item{alpha}{Numeric alpha.}

\item{margin}{Numeric error margin.}

\item{dependent.variable}{The variable being sampled as a vector.}

\item{weights}{Numeric weights (if present) as a vector.}
}
\value{
A list containing the parameters and a data.frame with the smoothed estimates.
}
\description{
Returns necessary sample size for a given power and dataset, but treats any standard deviation below 0.1 as 0.1
}
\keyword{internal}
