% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.functions.r
\name{sample.size}
\alias{sample.size}
\title{Sample size}
\usage{
sample.size(alpha, margin, dependent.variable, weights)
}
\arguments{
\item{alpha}{Numeric alpha.}

\item{margin}{Numeric error margin.}

\item{dependent.variable}{The variable being sampled as a vector.}

\item{weights}{Numeric weights (if present) as a vector.}
}
\value{
A list containing the parameters and a data.frame with the smoothed estimates.
}
\description{
Returns necessary sample size for a given power and dataset
}
\keyword{internal}
