% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d.smooths.r
\name{save.kernel.smooths}
\alias{save.kernel.smooths}
\title{Save kernel smooths in space and time}
\usage{
save.kernel.smooths(kernel.smooth, location, variable.name)
}
\arguments{
\item{kernel.smooth}{A list output of kernel.smooth.in.space.and.time() or kernel.smooth.in.space.and.time.with.margins().}

\item{location}{String location on the disk to save output to.}

\item{variable.name}{String name of the variable (used in filenames).}
}
\value{
No returned value, called to save data to disk.
}
\description{
Saves the output of kernel.smooth.in.space.and.time() or kernel.smooth.in.space.and.time.with.margins() to a directory
}
