% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.functions.r
\name{square.kernel}
\alias{square.kernel}
\title{Square kernel}
\usage{
square.kernel(d, bandwidth)
}
\arguments{
\item{d}{Numeric vector of distances.}

\item{bandwidth}{Numeric bandwidth.}
}
\value{
A numeric vector of kernel weights.
}
\description{
Square kernel function - weights all points within bandwidth as 1, all others as 0.
}
\keyword{internal}
