% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.functions.r
\name{triangular.kernel}
\alias{triangular.kernel}
\title{Triangular kernel}
\usage{
triangular.kernel(d, bandwidth)
}
\arguments{
\item{d}{Numeric vector of distances.}

\item{bandwidth}{Numeric bandwidth.}
}
\value{
A numeric vector of kernel weights.
}
\description{
Triangular kernel function - weights all points within bandwidth on a straight line such that where d=0 weight=1, and where d=bandwidth weight=0; all points beyond bandwidth are weighted 0.
}
\keyword{internal}
